add: content detail dialog skeleton
