--- src/components/ui/Content/hooks/useContentDetailDialog/component/index.tsx
@@ -1,7 +1,0 @@
-export type { ContentDetailDialogProps } from './ContentDetailDialog';
-import ContentDetailDialog from './ContentDetailDialog'
-import ContentDetailSkeleton from './ContentDetailSkeleton'
-
-export default Object.assign(ContentDetailDialog, {
-  Skeleton: ContentDetailSkeleton
-})
